requirements.txt
certifi==2018.8.24
chardet==3.0.4
cymem==1.31.2
cytoolz==0.9.0.1
dill==0.2.8.2
idna==2.7
msgpack==0.5.6
msgpack-numpy==0.4.3.1
murmurhash==0.28.0
numpy==1.15.1
Pillow==5.2.0
plac==0.9.6
preshed==1.0.1
regex==2017.4.5
requests==2.19.1
six==1.11.0
spacy==2.0.12
thinc==6.10.3
toolz==0.9.0
torchtext==0.2.3
torchvision==0.2.1
tqdm==4.25.0
ujson==1.35
urllib3==1.23
wrapt==1.10.11
